train/requirements.txt
torch>=1.13.0
torchvision>=0.14.0
diffusers>=0.23.0
transformers>=4.30.0
insightface>=0.7.3
ultralytics>=8.0.0
numpy>=1.22.0
matplotlib>=3.5.0
pillow>=9.0.0
tqdm>=4.64.0
wandb>=0.15.0
pyyaml>=6.0
scikit-learn>=1.0.0
tensorboard>=2.10.0 